+++ src/index.ts
@@ -28,7 +28,7 @@ export * from './assets/TokenRatesController';
 export * from './transaction/TransactionController';
 export * from './message-manager/PersonalMessageManager';
 export * from './message-manager/TypedMessageManager';
-export * from './notification/NotificationController';
+export * from './announcement/AnnouncementController';
 export * from './assets/TokenListController';
 export * from './gas/GasFeeController';
 export * from './assets/TokensController';